Add WebSocketValidator
