+++ src/Unosquare.Labs.EmbedIO/Modules/VirtualPaths.cs
@@ -106,6 +106,5 @@
             urlPath = urlPath.TrimStart(Path.DirectorySeparatorChar);
             return urlPath;
         }
-
     }
 }
